third_party/blink/web_tests/platform/generic/fast/events/mouse-double-triple-click-should-not-select-next-node-for-user-select-none-expected.txt

Mouse double/triple click on an element with "webkit-user-select none" should not select the next node

Foo
PASS doubleClick(); document.getSelection() is not "Foo"
PASS document.getSelection().removeAllRanges(); document.getSelection().toString().indexOf("Foo") is -1
PASS tripleClick(); document.getSelection().toString().indexOf("Foo") is -1
PASS successfullyParsed is true

TEST COMPLETE
